降低最低版本要求为 Android 4.0 (SDK 14)
